removed unused assests and moved the testimonial data into InfoTestimonial component
